requirements.txt
streamlit==1.40.2
ollama==0.4.4
pdfplumber==0.11.4
langchain==0.3.10
langchain-core==0.3.22
langchain-ollama==0.2.1
langchain_community==0.3.10
langchain_text_splitters==0.3.2
unstructured==0.16.11
unstructured[all-docs]
onnx==1.17.0
protobuf==3.20.3
chromadb==0.4.22
Pillow==11.0.0
numpy==1.26.4
IPython==8.30.0















